KP-11145 New report for survey statistics # Tests

+++ src/Core/BoundedContexts/Headquarters/WB.Core.BoundedContexts.Headquarters/Views/Reposts/SurveyStatistics/CategoricalReportViewItem.cs
@@ -9,7 +9,7 @@ namespace WB.Core.BoundedContexts.Headquarters.Views.Reposts.SurveyStatistics
         public string ResponsibleName { get; set; }
         public Guid TeamLeadId { get; set; }
         public string TeamLeadName { get; set; }
-        public int[] Values { get; set; }
+        public long[] Values { get; set; }
         public long Total => Values.Aggregate(0L, (l, i) => l + i);
     }
 }
